+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- ???
+- Added equivalent untis to `RexsStandardUnitIds` (`deg`, `degree`)
 
 
 ## [0.14.0] - 2025-07-18